Moved functions to utils
